#6594 use html file for modal template
